Added previews in event and group creation
