+++ package.json
@@ -61,6 +61,7 @@
     "sequelize": "^4.2.1",
     "serialize-javascript": "^1.3.0",
     "simple-line-icons": "^2.4.1",
+    "slugify": "^1.2.4",
     "source-map-support": "^0.5.0",
     "universal-router": "^4.3.0",
     "uuid": "^3.1.0",
