+++ src/routes/meetup/CreateGroup.js
@@ -2,87 +2,59 @@
 
 import React, { Component } from 'react';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
+import slugify from 'slugify';
 import {
   Container,
   Form,
-  Input,
   Header,
   TextArea,
   Button,
   Icon,
   Dropdown,
+  Modal,
 } from 'semantic-ui-react';
 import SemanticUI from '../../components/SemanticUI';
 import ProfilePictureUpload from '../../components/ProfilePictureUpload';
+import SingleGroup from './SingleGroup';
 import { log, logError } from '../../logger';
 import s from './CreateGroup.scss';
+import colors from './colors';
 
-const colorData = [
-  ['#FDEB71', '#F8D800'],
-  ['#ABDCFF', '#0396FF'],
-  ['#FEB692', '#EA5455'],
-  ['#CE9FFC', '#7367F0'],
-  ['#90F7EC', '#32CCBC'],
-  ['#FFF6B7', '#F6416C'],
-  ['#81FBB8', '#28C76F'],
-  ['#E2B0FF', '#9F44D3'],
-  ['#F97794', '#623AA2'],
-  ['#FCCF31', '#F55555'],
-  ['#F761A1', '#8C1BAB'],
-  ['#43CBFF', '#9708CC'],
-  ['#5EFCE8', '#736EFE'],
-  ['#FAD7A1', '#E96D71'],
-  ['#FFD26F', '#3677FF'],
-  ['#A0FE65', '#FA016D'],
-  ['#FFDB01', '#0E197D'],
-  ['#FEC163', '#DE4313'],
-  ['#92FFC0', '#002661'],
-  ['#EEAD92', '#6018DC'],
-  ['#F6CEEC', '#D939CD'],
-  ['#52E5E7', '#130CB7'],
-  ['#F1CA74', '#A64DB6'],
-  ['#E8D07A', '#5312D6'],
-  ['#EECE13', '#B210FF'],
-  ['#79F1A4', '#0E5CAD'],
-  ['#FDD819', '#E80505'],
-  ['#FFF3B0', '#CA26FF'],
-  ['#FFF5C3', '#9452A5'],
-  ['#F05F57', '#360940'],
-  ['#2AFADF', '#4C83FF'],
-  ['#FFF886', '#F072B6'],
-  ['#97ABFF', '#123597'],
-  ['#F5CBFF', '#C346C2'],
-  ['#FFF720', '#3CD500'],
-  ['#FF6FD8', '#3813C2'],
-  ['#EE9AE5', '#5961F9'],
-  ['#FFD3A5', '#FD6585'],
-  ['#C2FFD8', '#465EFB'],
-  ['#FD6585', '#0D25B9'],
-  ['#FD6E6A', '#FFC600'],
-  ['#65FDF0', '#1D6FA3'],
-  ['#6B73FF', '#000DFF'],
-  ['#FF7AF5', '#513162'],
-  ['#F0FF00', '#58CFFB'],
-  ['#FFE985', '#FA742B'],
-  ['#FFA6B7', '#1E2AD2'],
-  ['#FFAA85', '#B3315F'],
-  ['#72EDF2', '#5151E5'],
-  ['#FF9D6C', '#BB4E75'],
-  ['#F6D242', '#FF52E5'],
-  ['#69FF97', '#00E4FF'],
-  ['#3B2667', '#BC78EC'],
-  ['#70F570', '#49C628'],
-  ['#3C8CE7', '#00EAFF'],
-  ['#FAB2FF', '#1904E5'],
-  ['#81FFEF', '#F067B4'],
-  ['#FFA8A8', '#FCFF00'],
-  ['#FFCF71', '#2376DD'],
-  ['#FF96F9', '#C32BAC'],
-];
+const getCurrentMonth = () => {
+  switch (new Date().getMonth()) {
+    case 0:
+      return 'jan';
+    case 1:
+      return 'feb';
+    case 2:
+      return 'mar';
+    case 3:
+      return 'apr';
+    case 4:
+      return 'may';
+    case 5:
+      return 'jun';
+    case 6:
+      return 'jul';
+    case 7:
+      return 'aug';
+    case 8:
+      return 'sep';
+    case 9:
+      return 'oct';
+    case 10:
+      return 'nov';
+    case 11:
+      return 'dec';
+    default:
+      break;
+  }
+};
 
 class CreateGroup extends Component {
   state = {
     name: '',
+    slug: '',
     about: '',
     image: '',
     organizers: [],
@@ -92,6 +64,7 @@ class CreateGroup extends Component {
       { key: 'anna', text: 'Anna Sigurlaug' },
     ],
     selectedColor: '#FDEB71,#F8D800',
+    showPreview: false,
   };
   async handleSubmit() {
     const { email, password } = this.state;
@@ -122,10 +95,20 @@ class CreateGroup extends Component {
   }
   handleInputChange(key, data) {
     log('handleInputChange');
-    this.setState({
-      ...this.state,
-      [key]: data.value,
-    });
+    if (key === 'name') {
+      this.setState({
+        ...this.state,
+        name: data.value,
+        slug: slugify(data.value, {
+          lower: true,
+        }),
+      });
+    } else {
+      this.setState({
+        ...this.state,
+        [key]: data.value,
+      });
+    }
   }
   handleColorClicked(color) {
     this.setState({
@@ -148,6 +131,8 @@ class CreateGroup extends Component {
       organizers,
       selectedColor,
       organizersOptions,
+      showPreview,
+      slug,
     } = this.state;
     return (
       <SemanticUI>
@@ -156,19 +141,25 @@ class CreateGroup extends Component {
             Create Group
           </Header>
           <Form>
-            <Form.Field>
-              <label>Name</label>
-              <Input
+            <Form.Group>
+              <Form.Input
                 type="text"
                 value={name}
+                label="Name"
                 onChange={(e, d) => this.handleInputChange('name', d)}
               />
-            </Form.Field>
+              <Form.Input
+                type="text"
+                value={slug}
+                label="Slug"
+                onChange={(e, d) => this.handleInputChange('slug', d)}
+              />
+            </Form.Group>
             <Form.Field>
-              <label>About</label>
               <TextArea
                 autoHeight
                 value={about}
+                name="About"
                 style={{ minHeight: 100 }}
                 onChange={(e, d) => this.handleInputChange('about', d)}
               />
@@ -177,7 +168,7 @@ class CreateGroup extends Component {
             <Form.Field>
               <label>Background</label>
               <div className={s.colors}>
-                {colorData.map(color => (
+                {colors.map(color => (
                   <div
                     key={color.toString()}
                     className={s.color}
@@ -231,11 +222,56 @@ class CreateGroup extends Component {
               />
             </Form.Field> */}
 
+            <Button
+              size="large"
+              onClick={() => this.setState({ showPreview: true })}
+            >
+              Preview
+            </Button>
             <Button primary size="large" onClick={() => this.handleSubmit()}>
               Create group
             </Button>
           </Form>
         </Container>
+        <Modal
+          basic
+          closeOnDocumentClick
+          size="fullscreen"
+          open={showPreview}
+          onClose={() => this.setState({ showPreview: false })}
+        >
+          <Header content="Preview" />
+          <Modal.Content
+            style={{
+              background: '#fff',
+            }}
+          >
+            <SingleGroup
+              backgroundColor={selectedColor.split(',')}
+              name={name}
+              about={about}
+              logo={image ? `https://communityfund.imgix.net/${image}` : null}
+              events={[
+                {
+                  day: new Date().getDate(),
+                  month: getCurrentMonth(),
+                  title: 'Our First awesome event',
+                  shortDescription: `NOT VISIBLE ON GROUP PROFILE! We are super excited to announce the first meetup of ${name}. Come and join us for a evening of fantasting talks and discussion.`,
+                },
+              ]}
+            />
+          </Modal.Content>
+          <Modal.Actions>
+            <Button
+              basic
+              color="red"
+              inverted
+              onClick={() => this.setState({ showPreview: false })}
+            >
+              <Icon name="remove" /> Close
+            </Button>
+          </Modal.Actions>
+        </Modal>
       </SemanticUI>
     );
   }
