+++ src/routes/meetup/SingleEvent.scss
@@ -1,6 +1,6 @@
 .header {
   height: 100vh;
-  width: 100vw;
+  width: 100%;
   min-height: 400px;
 }
 
@@ -12,7 +12,9 @@
   align-items: center;
 }
 
-.text {
+.title {
+  display: flex;
+  flex-shrink: 0;
   font-weight: 'normal';
   margin-bottom: 0;
   text-transform: 'uppercase';
@@ -26,10 +28,19 @@
 .logo {
   // box-shadow: 3px 3px 20px rgba(0, 0, 0, 0.2);
   // background: #f9df36 !important;
+  display: flex;
+  flex-shrink: 0;
+}
+
+.attendButton {
+  display: flex !important;
+  flex-shrink: 0;
+  margin-top: 10px !important;
 }
 
 .eventDescription {
   font-size: 1.2em;
   font-weight: normal;
   font-family: helvetica;
+  color: #000;
 }
