+++ src/routes/meetup/SingleGroup.scss
@@ -1,6 +1,6 @@
 .header {
   height: 50vh;
-  width: 100vw;
+  width: 100%;
   min-height: 330px;
 }
