--- src/routes/meetup/attendees.js
@@ -1,63 +1,0 @@
-export default [
-  {
-    id: '2d130080-cc66-11e7-8744-09b6d394801c',
-    name: 'Reviewer 0',
-    title: 'CEO of awesome corp',
-    image:
-      'https://communityfund.imgix.net/e5351d4d-f7fc-46a2-ade8-cdde3303fa97.png?fit=crop&w=500&h=500',
-    topics: [],
-  },
-  {
-    id: '2d132790-cc66-11e7-8744-09b6d394801c',
-    name: 'Reviewer 1',
-    title: 'CEO of awesome corp',
-    image:
-      'https://communityfund.imgix.net/e5351d4d-f7fc-46a2-ade8-cdde3303fa97.png?fit=crop&w=500&h=500',
-    topics: [],
-  },
-  {
-    id: '2d132791-cc66-11e7-8744-09b6d394801c',
-    name: 'Reviewer 2',
-    title: 'CEO of awesome corp',
-    image:
-      'https://communityfund.imgix.net/e5351d4d-f7fc-46a2-ade8-cdde3303fa97.png?fit=crop&w=500&h=500',
-    topics: [],
-  },
-  {
-    id: '2d132793-cc66-11e7-8744-09b6d394801c',
-    name: 'Reviewer 4',
-    title: 'CEO of awesome corp',
-    image:
-      'https://communityfund.imgix.net/e5351d4d-f7fc-46a2-ade8-cdde3303fa97.png?fit=crop&w=500&h=500',
-    topics: [],
-  },
-  {
-    id: '2d132792-cc66-11e7-8744-09b6d394801c',
-    name: 'Reviewer 3',
-    title: 'CEO of awesome corp',
-    image:
-      'https://communityfund.imgix.net/e5351d4d-f7fc-46a2-ade8-cdde3303fa97.png?fit=crop&w=500&h=500',
-    topics: [],
-  },
-  {
-    id: '2d257710-cc66-11e7-8744-09b6d394801c',
-    name: 'Operator 1',
-    title: 'CEO of awesome corp',
-    image:
-      'https://communityfund.imgix.net/e5351d4d-f7fc-46a2-ade8-cdde3303fa97.png?fit=crop&w=500&h=500',
-    topics: [
-      {
-        name: 'ux',
-        color: 'blue',
-      },
-      {
-        name: 'development',
-        color: 'green',
-      },
-      {
-        name: 'art',
-        color: 'red',
-      },
-    ],
-  },
-];
